requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.1.1
werkzeug==2.3.7
python-docx==1.0.1
openai==1.61.0
python-dotenv==1.0.0
authlib==1.2.1
requests==2.31.0
flask-oauthlib==0.9.6
Flask-Session==0.5.0
nltk==3.8.1
numpy==1.24.3
# Chỉ định phiên bản cụ thể của scipy để tương thích với gensim
scipy==1.10.1
scikit-learn==1.2.2
gensim==4.3.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
# Các thư viện đã được thêm để khắc phục lỗi ModuleNotFoundError

